Move Pos to common
