+++ src/common/mod.rs
@@ -4,7 +4,7 @@ mod testing;
 mod vartable;
 
 pub use io::*;
-pub use source::{Source, SourceId, SourceMap};
+pub use source::{Pos, Source, SourceId, SourceMap};
 pub use testing::*;
 pub use vartable::VarTable;
