+++ src/scanner/mod.rs
@@ -4,8 +4,8 @@ mod scanner_test;
 use tracing::{debug, info, trace};
 
 use crate::{
-    ast::{Pos, Token, TokenKind, str_to_token},
-    common::Source,
+    ast::{Token, TokenKind, str_to_token},
+    common::{Pos, Source},
     config::Config,
     error::{Diagnostics, Report, Res},
 };
